Update mahmi.mgloader to use new database/common API
mahmi-mgloader has been fixed to use the new API of mahmi-common and
mahmi-database that does not use Validation types

Also, some minnor fixes in mahmi-common and mahmi-database.

+++ mahmi-common/src/main/java/es/uvigo/ei/sing/mahmi/common/entities/fasta/Fasta.java
@@ -14,6 +14,6 @@ import es.uvigo.ei.sing.mahmi.common.entities.sequences.ChemicalCompoundSequence
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public abstract class Fasta<A extends ChemicalCompoundSequence<? extends ChemicalCompound>> {
 
-    protected final Map<A, Integer> sequences;
+    protected final Map<A, Long> sequences;
 
 }
