+++ mahmi-common/src/main/java/es/uvigo/ei/sing/mahmi/common/serializers/jaxb/entities/fasta/ProteinFastaAdapter.java
@@ -9,15 +9,15 @@ import javax.xml.bind.annotation.adapters.XmlAdapter;
 import es.uvigo.ei.sing.mahmi.common.entities.fasta.ProteinFasta;
 import es.uvigo.ei.sing.mahmi.common.entities.sequences.AminoAcidSequence;
 
-public final class ProteinFastaAdapter extends XmlAdapter<ProteinFasta, Map<String, Integer>> {
+public final class ProteinFastaAdapter extends XmlAdapter<ProteinFasta, Map<String, Long>> {
 
     @Override
-    public Map<String, Integer> unmarshal(final ProteinFasta fasta) throws Exception {
+    public Map<String, Long> unmarshal(final ProteinFasta fasta) throws Exception {
         return mapKeys(fasta.getSequences(), AminoAcidSequence::toString);
     }
 
     @Override
-    public ProteinFasta marshal(final Map<String, Integer> map) throws Exception {
+    public ProteinFasta marshal(final Map<String, Long> map) throws Exception {
         return ProteinFasta.of(mapKeys(map, AminoAcidSequence::fromString));
     }
